+++ src/main/java/org/dbos/apiary/benchmarks/retro/WordPressBenchmark.java
@@ -1,0 +1,16 @@
+package org.dbos.apiary.benchmarks.retro;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.List;
+
+public class WordPressBenchmark {
+    private static final Logger logger = LoggerFactory.getLogger(WordPressBenchmark.class);
+
+    public static void benchmark(String dbAddr, Integer interval, Integer duration, boolean skipLoad, int retroMode, long startExecId, long endExecId, String bugFix, List<Integer> percentages) {
+        // TODO: implement this.
+        logger.info("Implementing wordpress bench.");
+
+    }
+}
